Move colormap and splatmap sampling to the fragment shader.
This is to workaround an unexpected blow-up of interpolated values when
they were sampled in the vertex shader and used in the fragment shader.
The assumption is that interpolation somehow runs out of precision or
incorrectly calculates varying values, causing them to become higher than
1, which in turn causes flashes to occur when bloom is enabled.
I think this is still unexpected and needs more investigation,
but I'm not sure what else to try (that I know how to do)
and I don't have much time.
See https://github.com/Zylann/godot_heightmap_plugin/issues/469